requirements.txt
longport>=3.0.22
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
pydantic>=2.0.0
python-dotenv>=1.0.0
